Handle union ordering in modules pipeline
